fix: use pinned version for alpine (#153)
Make does not work with alpine 3.14, this commit pins the alpine version
to 3.13 to fix the issue. The root cause seems to be [this issue][1] with
Alpine and Docker 20.10.5 and earlier. How this issue presents itself in
other container runtimes is unclear, but consensus seems to be that
reverting to alpine 3.13 fixes the issue.

[1] https://github.com/alpinelinux/docker-alpine/issues/182

Fixes #152